cleanup mapping root resolution logic
